Déplacement des models de données dans Shared
